requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.23.0
attrs==17.2.0
autobahn==17.9.3
Automat==0.6.0
cffi==1.11.0
channels==1.1.8
constantly==15.1.0
cryptography==2.0.3
daphne==1.3.0
dj-database-url==0.4.2
Django==1.11.5
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.4
djangorestframework-filters==0.10.1
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
h2==3.0.1
hpack==3.0.0
hyperframe==5.1.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
Markdown==2.6.9
msgpack-python==0.4.8
priority==1.3.0
psycopg2==2.7.3.1
pyasn1==0.3.6
pyasn1-modules==0.1.4
pycparser==2.18
PyJWT==1.5.3
pyOpenSSL==17.3.0
pytz==2017.2
redis==2.10.6
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
zope.interface==4.4.3
